requirements.txt
absl-py==2.0.0
altair==5.3.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
billiard==4.2.1
blinker==1.7.0
cachetools==5.3.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==5.1
dnspython==2.6.1
docx2pdf==0.1.8
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.5
filelock==3.15.4
flatbuffers==23.5.26
fonttools==4.53.1
fsspec==2024.10.0
fuzzywuzzy==0.18.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.59.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.2
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.7
imageio==2.35.1
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.14.0
kiwisolver==1.4.5
kombu==5.4.2
lazy_loader==0.3
Levenshtein==0.25.1
libclang==16.0.6
lxml==5.3.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.2
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.3
phonenumbers==8.13.40
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.0
priority==2.0.0
prompt-toolkit==3.0.43
protobuf==4.24.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.13.5
pycountry==24.6.1
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.8.1b0
Pygments==2.18.0
PyJWT==2.8.0
pypandoc==1.13
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
rapidfuzz==3.9.1
redis==5.1.1
referencing==0.34.0
regex==2024.5.15
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.25
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tf-slim==1.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.20.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.46.1
typer==0.12.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
Unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.30.1
vine==5.1.0
virtualenv==20.26.3
watchdog==4.0.0
watchfiles==0.23.0
wcwidth==0.2.13
websockets==13.0
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
xlwings==0.31.10
